DELIA-27970: Avoid multiple restart of systemd timesyncd
Reason for change: The systemd timesyncd service has to be restarted only once as soon as the default route is available.
The nlmon calls the subscribed script for default routes from all connected gateway and XB devices on the network.
Test Procedure: Validate channel tune is working on rogers environment
Risks: low

Change-Id: I9131457d88c9843f359eaaebb4cbab896668a163
Signed-off-by: mselvaraj <mani_selvaraj@comcast.com>